refactor(store): move attachments and migrations to separate modules
move attachment-related code to attachments.rs and migrations to migrations.rs
clean up store/mod.rs by removing migrated code
update schema.sql with latest view definitions
