+++ src/store/inbox.rs
@@ -89,89 +89,6 @@ impl InboxConversationStateView {
     }
 }
 
-pub(super) fn refresh_inbox_conversation_state_view(conn: &Connection) -> Result<()> {
-    conn.execute_batch(
-        "
-        DROP VIEW IF EXISTS inbox_conversation_state_view;
-        CREATE VIEW inbox_conversation_state_view AS
-        SELECT
-            cm.channel_id AS conversation_id,
-            c.name AS conversation_name,
-            c.channel_type AS conversation_type,
-            cm.member_name AS member_name,
-            cm.member_type AS member_type,
-            COALESCE(irs.last_read_seq, 0) AS last_read_seq,
-            irs.last_read_message_id AS last_read_message_id,
-            (
-                SELECT COUNT(*)
-                FROM messages top_level
-                WHERE top_level.channel_id = cm.channel_id
-                  AND top_level.thread_parent_id IS NULL
-                  AND top_level.seq > COALESCE(irs.last_read_seq, 0)
-            ) + (
-                SELECT COUNT(*)
-                FROM messages reply
-                LEFT JOIN inbox_thread_read_state itrs
-                  ON itrs.conversation_id = reply.channel_id
-                 AND itrs.thread_parent_id = reply.thread_parent_id
-                 AND itrs.member_name = cm.member_name
-                WHERE reply.channel_id = cm.channel_id
-                  AND reply.thread_parent_id IS NOT NULL
-                  AND reply.seq > COALESCE(itrs.last_read_seq, 0)
-                  AND (
-                    cm.member_type != 'agent'
-                    OR EXISTS (
-                        SELECT 1
-                        FROM messages parent
-                        WHERE parent.id = reply.thread_parent_id
-                          AND parent.channel_id = cm.channel_id
-                          AND parent.sender_type = 'agent'
-                          AND parent.sender_name = cm.member_name
-                    )
-                    OR EXISTS (
-                        SELECT 1
-                        FROM messages prior
-                        WHERE prior.channel_id = cm.channel_id
-                          AND prior.thread_parent_id = reply.thread_parent_id
-                          AND prior.sender_type = 'agent'
-                          AND prior.sender_name = cm.member_name
-                          AND prior.seq < reply.seq
-                    )
-                  )
-            ) AS unread_count
-        FROM channel_members cm
-        JOIN channels c ON c.id = cm.channel_id
-        LEFT JOIN inbox_read_state irs
-          ON irs.conversation_id = cm.channel_id
-         AND irs.member_name = cm.member_name;
-        ",
-    )?;
-    Ok(())
-}
-
-pub(super) fn migrate_inbox_read_state(conn: &Connection) -> Result<()> {
-    conn.execute(
-        "INSERT OR IGNORE INTO inbox_read_state (
-            conversation_id, member_name, member_type, last_read_seq, last_read_message_id
-         )
-         SELECT
-            cm.channel_id,
-            cm.member_name,
-            cm.member_type,
-            cm.last_read_seq,
-            (
-                SELECT m.id
-                FROM messages m
-                WHERE m.channel_id = cm.channel_id
-                  AND m.seq = cm.last_read_seq
-                LIMIT 1
-            )
-         FROM channel_members cm",
-        [],
-    )?;
-    Ok(())
-}
-
 impl Store {
     fn latest_conversation_message_inner(
         conn: &Connection,
